Fortify more chrome extension styling
