+++ src/chrome-extension/content-script/components/save-page-message/components/save-page-content/save-note-editor/SaveNoteEditor.tsx
@@ -1,16 +1,12 @@
 import "./save-note-editor.scss"
 import React from "react";
-import {Slate, ReactEditor} from "slate-react";
 import {SlatePlugins} from "@udecode/slate-plugins";
 import {Node} from "slate";
 import {equals} from "ramda";
-import {defaultComponents} from "../../../../../../../common/rich-text-editor/components";
 import {defaultOptions} from "../../../../../../../common/rich-text-editor/options";
-import {
-    defaultEditableProps,
-    useBasicPlugins,
-    usePeakPlugins
-} from "../../../../../../../common/rich-text-editor/editorFactory";
+import { defaultEditableProps } from "../../../../../../../common/rich-text-editor/editorFactory";
+import {useBasicPlugins} from "../../../../../../../common/rich-text-editor/base_plugins";
+import {useComponents} from "../../../../../../../common/rich-text-editor/components";
 
 export const SaveNoteEditor = (props: { initialValue: Node[], onChange: (newValue: Node[]) => void }) => {
     const { initialValue, onChange } = props
@@ -25,7 +21,7 @@ export const SaveNoteEditor = (props: { initialValue: Node[], onChange: (newValu
         <SlatePlugins
             id={"extensionNoteEditor"}
             plugins={useBasicPlugins()}
-            components={defaultComponents}
+            components={useComponents(false, false)}
             onChange={updatePageContent}
             options={defaultOptions}
             editableProps={defaultEditableProps}
