+++ src/chrome-extension/content-script/components/save-page-message/components/save-page-content/save-page-content.scss
@@ -11,6 +11,57 @@ $drawer-title-height: 45px;
   flex-grow: 1;
   width: 100%;
 
+  span {
+    font-family: "Inter"!important;
+    color: $text-color!important;
+  }
+  p {
+    font-family: "Inter"!important;
+    color: $text-color!important;
+  }
+
+  h1, h2, h3, h4, h5 {
+    font-family: "Inter"!important;
+    color: $heading-color!important;
+    font-variant: normal!important;
+    letter-spacing: -.025rem!important;
+    margin-top: 1.2em!important;
+    margin-bottom: .5em!important;
+    //margin-top: 1.5em!important;
+  }
+
+  h1 {
+    font-size: 2.0rem!important;
+    font-weight: 800!important;
+  }
+
+  h2 {
+    font-size: 1.65rem!important; // 1.5?
+    font-weight: 700!important;
+    line-height: 1.3333;
+  }
+
+  h3 {
+    font-size: 1.35em!important; // 1.25?
+    font-weight: 600!important;
+  }
+
+  h4 {
+    font-size: 1.1rem!important;
+    font-weight: 600!important;
+  }
+
+  h5 {
+    font-size: 1.0rem!important;
+    font-weight: 600!important;
+  }
+
+  body::placeholder {
+    color: #8992a4!important;
+  }
+
+
+
   // Necessary for Youtube
   background-color: white!important;
 
@@ -44,6 +95,7 @@ $drawer-title-height: 45px;
       }
 
       .page-peak-title-input {
+        font-weight: 400!important;
         position: relative!important;
         font-family: $font-family!important;
         width: 100%;
@@ -119,6 +171,8 @@ $drawer-title-height: 45px;
     }
 
     &.footer {
+      display: flex!important;
+      flex-direction: row!important;
       //margin: -$drawer-padding-amount;
       //border-top: 1px solid #f0f0f0;
       width: 100% !important;
@@ -128,6 +182,7 @@ $drawer-title-height: 45px;
       font-size: 18px;
       height: 126px!important;
       background-color: white!important;
+      background-image: none!important;
 
       justify-content: center;
       align-items: center;
