+++ src/chrome-extension/content-script/components/save-page-message/components/save-page-header-content/saved-page-header-content.scss
@@ -16,9 +16,10 @@
     background-color: white!important;
 
     .peak-message-header {
+      font-family: "Inter"!important;
+      color: $text-color!important;
       font-size: 20px!important;
       font-weight: 600!important;
-      color: $text-color!important;
       margin-bottom: 0px!important;
       margin-top: 0px!important;
       letter-spacing: normal!important;
