--- src/common/rich-text-editor/editors/chrome-extension/config.tsx
@@ -1,4 +1,0 @@
-import { SlatePlugin } from "@udecode/slate-plugins";
-import {usePeakPlugins} from "../../editorFactory";
-
-// export const extensionPlugins: SlatePlugin[] = usePeakPlugins()
